% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmkin.R
\name{gmkin}
\alias{gmkin}
\title{Start a graphical user interface (GUI) based on the \code{gWidgetsWWW2} toolkit.}
\usage{
gmkin(script_name, show.log = FALSE)
}
\arguments{
\item{script_name}{During development, a script name with a local working
version of gmkin can be passed. Defaults to the location of the gmkin.R
script shipped with the package.}

\item{show.log}{During development, it may be useful to see the log of the
Rook apps.}
}
\value{
The function is called for its side effect, namely starting the GUI
  in a browser. For the curious, the desperate or the adventurous, the gmkin
  app (a GWidgetsApp object) is returned.
}
\description{
This function starts a browser based GUI. Please visit the 
\href{http://github.com/jverzani/gWidgetsWWW2}{github page of gWidgetsWWW2} 
for an explanation how this toolkit works.
}
